+++ src/model/Discountable.java
@@ -2,5 +2,5 @@ package model;
 
 
 public interface Discountable {
-    int getDiscount();
+    double getDiscount();
 }
